Simplify topo sorted traversal to depth first traversal
Each node points to 0 or 1 other node, never to more. There's only one
unique path by which each node can be reached by going backwards from
the output node. This means that in the `backward()` method `v` is never
already in `visited`. The algorithm thus simplifies to a depth-first
traversal.

Side-note: if a node could be reached by more than 1 path from the output
node, the node would have a different gradient for each of the paths. So,
while you could reuse nodes for the forward calculation and get correct
results, there's no facility to store several gradients per node. Thus
it's unnecessary to account for this possibility here.